requirements.txt
dnspython
python-whois
requests